Dude it works.
Welder now has an unshaded flame toggle!

Door component no longer on client!
