+++ Content.Client/Content.Client.csproj
@@ -72,7 +72,6 @@
     <Compile Include="GameObjects\Components\Storage\ClientStorageComponent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="GameObjects\Components\Items\ClientHandsComponent.cs" />
-    <Compile Include="GameObjects\Components\Doors\ClientDoorComponent.cs" />
     <Compile Include="Interfaces\GameObjects\Components\Items\IHandsComponent.cs" />
     <Compile Include="UserInterface\HandsGui.cs" />
   </ItemGroup>
